+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Google Analytics. ([#45](https://github.com/imse-ty/imsety.com/issues/45))
+- Changelog page for public release notes. [[#47](https://github.com/imse-ty/imsety.com/issues/47)]
 
 ## [0.6.0] - 2022-01-29