+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Added
 
 - Google Analytics. ([#45](https://github.com/imse-ty/imsety.com/issues/45))
-- Changelog page for public release notes. [[#47](https://github.com/imse-ty/imsety.com/issues/47)]
+- Changelog page for public release notes. ([#47](https://github.com/imse-ty/imsety.com/issues/47))
 
 ### Changed