+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Changed
 
 - Footer copyright notice thing has been updated from "2021" to "2022." ([#49](https://github.com/imse-ty/imsety.com/issues/49))
+- Upgrade to Next.js 12. ([#53](https://github.com/imse-ty/imsety.com/issues/53))
 
 ## [0.6.0] - 2022-01-29