+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Upgraded to Next.js 12. ([#53](https://github.com/imse-ty/imsety.com/issues/53))
 - Projects on the home page are now sorted by date published. ([#52](https://github.com/imse-ty/imsety.com/issues/52))
 - Made sure project entry info such as year, timeline, tools, and summary are hidden when not in use.
+- `title` and `slug` for project entries and blog posts are now required.
 
 ## [0.6.0] - 2022-01-29