+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+## [1.0.0] - 2022-02-19
+
+First stable release 🥳🥳🎉
+
 ### Added
 
 - Google Analytics. ([#45](https://github.com/imse-ty/imsety.com/issues/45))
@@ -98,7 +102,8 @@
 
 Initial development.
 
-[unreleased]: https://github.com/imse-ty/imsety.com/compare/v0.6.0...HEAD
+[unreleased]: https://github.com/imse-ty/imsety.com/compare/v1.0.0...HEAD
+[1.0.0]: https://github.com/imse-ty/imsety.com/compare/v0.6.0...v1.0.0
 [0.6.0]: https://github.com/imse-ty/imsety.com/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/imse-ty/imsety.com/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/imse-ty/imsety.com/compare/v0.3.0...v0.4.0